Refactor mainview and mainviewpresenter
Utilized the explicit operator to cast a DataGridViewRow to a Password
object. Now passing the DataGridViewRow between events instead of all
of the text fields. This should make it much easier to add new
properties to Password since only the explicit operator needs to be
updated.

Should do something similar for the text box fields, maybe using
binding to a temporary view model that can be passed as event instead
of multiple strings.

IS8_32_AddEmailProperty